refactor: DB dto 패키지 이동
