+++ src/main/java/chess/dto/db/MovementResponse.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.db;
 
 import chess.domain.game.Movement;
